Add one test for validateFields
